Move some branch options to magit-git.el
This is a bit unfortunate, but we have already done the same for
other branch options in the past and this helps avoiding circular
dependencies.  The commands whose behavior depends on these options
are mostly defined in magit-branch, but the actual logic is
implemented in functions that are defined in magit-git.el, which is
loaded much earlier and cannot depend on magit-branch, not least
because magit-branch, like most libraries, depends on magit-git.